+++ my/core/common.py
@@ -5,7 +5,6 @@ from dataclasses import is_dataclass, asdict as dataclasses_asdict
 import functools
 from contextlib import contextmanager
 import os
-import sys
 from typing import (
     Any,
     Callable,
@@ -380,45 +379,6 @@ def assert_subpackage(name: str) -> None:
     # can lead to some unexpected issues if you 'import cachew' which being in my/core directory.. so let's protect against it
     # NOTE: if we use overlay, name can be smth like my.origg.my.core.cachew ...
     assert name == '__main__' or 'my.core' in name, f'Expected module __name__ ({name}) to be __main__ or start with my.core'
-
-
-from .compat import ParamSpec
-_P = ParamSpec('_P')
-_T = TypeVar('_T')
-
-# https://stackoverflow.com/a/10436851/706389
-from concurrent.futures import Future, Executor
-class DummyExecutor(Executor):
-    def __init__(self, max_workers: Optional[int]=1) -> None:
-        self._shutdown = False
-        self._max_workers = max_workers
-
-    if TYPE_CHECKING:
-        if sys.version_info[:2] <= (3, 8):
-            # 3.8 doesn't support ParamSpec as Callable arg :(
-            # and any attempt to type results in incompatible supertype.. so whatever
-            def submit(self, fn, *args, **kwargs): ...
-        else:
-            def submit(self, fn: Callable[_P, _T], /, *args: _P.args, **kwargs: _P.kwargs) -> Future[_T]: ...
-    else:
-        def submit(self, fn, *args, **kwargs):
-            if self._shutdown:
-                raise RuntimeError('cannot schedule new futures after shutdown')
-
-            f: Future[Any] = Future()
-            try:
-                result = fn(*args, **kwargs)
-            except KeyboardInterrupt:
-                raise
-            except BaseException as e:
-                f.set_exception(e)
-            else:
-                f.set_result(result)
-
-            return f
-
-    def shutdown(self, wait: bool=True, **kwargs) -> None:
-        self._shutdown = True
 
 
 # TODO deprecate and suggest to use one from my.core directly? not sure
